# students/k3341/Moskalets_Danila/Lr1/teamfinder-api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: TeamFinder API
  description: Платформа для поиска людей в команду для совместной работы над проектами
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags: [auth]
      summary: Вход в систему (получение JWT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Обновить access-токен
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  description: Refresh-токен из предыдущего ответа
      responses:
        '200':
          description: Новая пара токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [auth]
      summary: Получить текущего пользователя
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Данные текущего пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─── USERS ───────────────────────────────────────────────
...
  /users/me/skills:
    get:
      tags: [users]
      summary: Получить навыки текущего пользователя
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список навыков
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSkillResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [users]
      summary: Добавить навык текущему пользователю
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserSkillRequest'
      responses:
        '201':
          description: Навык добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSkillResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Навык уже добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me/skills/{skill_id}:
    delete:
      tags: [users]
      summary: Удалить навык у текущего пользователя
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SkillId'
      responses:
        '204':
          description: Навык удалён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── SKILLS ──────────────────────────────────────────────
  /skills:
    get:
      tags: [skills]
      summary: Получить список всех навыков
      parameters:
        - name: category
          in: query
          description: Фильтр по категории
          schema:
            type: string
        - name: search
          in: query
          description: Поиск по названию навыка
          schema:
            type: string
      responses:
        '200':
          description: Список навыков
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SkillResponse'

    post:
      tags: [skills]
      summary: Создать навык (только ADMIN)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSkillRequest'
      responses:
        '201':
          description: Навык создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Навык с таким названием уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── PROJECTS ────────────────────────────────────────────
  /projects:
    get:
      tags: [projects]
      summary: Поиск и фильтрация проектов
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProjectStatus'
        - name: skill_ids
          in: query
          description: Фильтр по ID требуемых навыков
          schema:
            type: array
            items:
              type: integer
        - name: owner_id
          in: query
          description: Фильтр по владельцу проекта
          schema:
            type: integer
        - name: search
          in: query
          description: Поиск по названию или описанию
          schema:
            type: string
        - name: sort_by
          in: query
          description: Поле для сортировки
          schema:
            type: string
            enum: [created_at, deadline, title]
            default: created_at
        - name: sort_order
          in: query
          description: Направление сортировки
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          description: Номер страницы (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Количество записей на странице
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Список проектов с пагинацией
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

    post:
      tags: [projects]
      summary: Создать проект
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Проект создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /projects/{project_id}:
    get:
      tags: [projects]
      summary: Получить проект по ID
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Проект
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [projects]
      summary: Обновить проект (только владелец)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Обновлённый проект
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [projects]
      summary: Удалить проект (только владелец или ADMIN)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '204':
          description: Проект удалён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{project_id}/skills:
    get:
      tags: [projects]
      summary: Получить все требуемые навыки проекта
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Список требуемых навыков
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectSkillResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [projects]
      summary: Добавить требуемый навык к проекту
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectSkillRequest'
      responses:
        '201':
          description: Навык добавлен к проекту
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSkillResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{project_id}/skills/{skill_id}:
    delete:
      tags: [projects]
      summary: Удалить навык из проекта
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/SkillId'
      responses:
        '204':
          description: Навык удалён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── TEAMS ───────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SkillId:
      name: skill_id
      in: path
      required: true
      schema:
        type: integer
      description: ID навыка
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: integer
      description: ID проекта
...
  responses:
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    UserRole:
      type: string
      enum: [ADMIN, USER]
      description: Роль пользователя
    ProficiencyLevel:
      type: string
      enum: [BEGINNER, INTERMEDIATE, ADVANCED, EXPERT]
      description: Уровень владения навыком
    ProjectStatus:
      type: string
      enum: [DRAFT, OPEN, IN_PROGRESS, COMPLETED, ARCHIVED]
      description: Статус проекта
...
    SkillImportance:
      type: string
      enum: [NICE_TO_HAVE, IMPORTANT, REQUIRED]
      description: Важность навыка для проекта
...
    Pagination:
      type: object
      required: [total, page, page_size, total_pages]
      properties:
        total:
          type: integer
          description: Общее количество записей
          example: 150
        page:
          type: integer
          description: Текущая страница (1-based)
          minimum: 1
          example: 2
        page_size:
          type: integer
          description: Размер страницы
          minimum: 1
          maximum: 100
          example: 20
        total_pages:
          type: integer
          description: Всего страниц
          example: 8

    # ── AUTH ──
    RegisterRequest:
      type: object
      required: [first_name, last_name, email, password]
      properties:
        first_name:
          type: string
          maxLength: 64
          example: John
        last_name:
          type: string
          maxLength: 64
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          minLength: 8
          example: securepassword123
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          example: securepassword123
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          example: bearer

    # ── USERS ──
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        role:
          $ref: '#/components/schemas/UserRole'
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        email:
          type: string
          example: john.doe@example.com
        avatar_url:
          type: string
          nullable: true
          example: https://example.com/avatar.jpg
        bio:
          type: string
          nullable: true
          example: Passionate developer
        is_verified:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
...
    SkillResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Python
        category:
          type: string
          example: Programming
    CreateSkillRequest:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
          maxLength: 128
          example: FastAPI
        category:
          type: string
          maxLength: 64
          example: Backend
    UserSkillResponse:
      type: object
      properties:
        skill:
          $ref: '#/components/schemas/SkillResponse'
        proficiency_level:
          $ref: '#/components/schemas/ProficiencyLevel'
    AddUserSkillRequest:
      type: object
      required: [skill_id, proficiency_level]
      properties:
        skill_id:
          type: integer
          example: 1
        proficiency_level:
          $ref: '#/components/schemas/ProficiencyLevel'

    # ── PROJECTS ──
    ProjectResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        owner_id:
          type: integer
          example: 1
        title:
          type: string
          example: E-commerce Platform
        description:
          type: string
          example: Building a modern e-commerce platform
        status:
          $ref: '#/components/schemas/ProjectStatus'
        deadline:
          type: string
          format: date
          nullable: true
          example: 2024-12-31
        created_at:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2024-01-15T00:00:00Z
    ProjectDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ProjectResponse'
        - type: object
          properties:
            required_skills:
              type: array
              items:
                $ref: '#/components/schemas/ProjectSkillResponse'
            teams:
              type: array
              items:
                $ref: '#/components/schemas/TeamResponse'
    ProjectListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateProjectRequest:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          maxLength: 256
          example: AI Chatbot
        description:
          type: string
          example: Developing an AI-powered chatbot
        deadline:
          type: string
          format: date
          nullable: true
          example: 2024-06-30
    UpdateProjectRequest:
      type: object
      properties:
        title:
          type: string
          example: AI Chatbot v2
        description:
          type: string
          example: Updated description
        status:
          $ref: '#/components/schemas/ProjectStatus'
        deadline:
          type: string
          format: date
          nullable: true
          example: 2024-07-31
    ProjectSkillResponse:
      type: object
      properties:
        skill:
          $ref: '#/components/schemas/SkillResponse'
        importance:
          $ref: '#/components/schemas/SkillImportance'
    AddProjectSkillRequest:
      type: object
      required: [skill_id, importance]
      properties:
        skill_id:
          type: integer
          example: 1
        importance:
          $ref: '#/components/schemas/SkillImportance'

    # ── TEAMS ──
    TeamResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        project_id:
          type: integer
          example: 1
        created_by:
          type: integer
          example: 1
        name:
          type: string
          example: Backend Team
        description:
          type: string
          nullable: true
          example: Responsible for API development
        created_at:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: integer
              example: 404
            message:
              type: string
              example: Resource not found
            details:
              type: object
              description: Дополнительные сведения об ошибке
              additionalProperties: true
              nullable: true
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: integer
              example: 422
            message:
              type: string
              example: Validation error
            details:
              type: array
              items:
                type: object
                properties:
                  loc:
                    type: array
                    items:
                      type: string
                    example: ["body", "email"]
                  msg:
                    type: string
                    example: field required
                  type:
                    type: string
                    example: missing